feat: add heritance and split modules
