+++ .pre-commit-config.yaml
@@ -1,10 +1,10 @@
 repos:
   - repo: https://github.com/psf/black
-    rev: 24.2.0  # versão estável mais recente do Black
+    rev: 24.2.0
     hooks:
       - id: black
         language_version: python3
-        args: [--line-length=88]  # padrão do Black
+        args: [--line-length=88]
 
   - repo: https://github.com/pycqa/isort
     rev: 5.13.2
@@ -18,7 +18,7 @@ repos:
       - id: flake8
         args: [
           "--max-line-length=88",
-          "--ignore=E203,W503" # compatibilidade com black
+          "--ignore=E203,W503"
         ]
 
   - repo: https://github.com/pre-commit/pre-commit-hooks
